+++ app/utils.py
@@ -1,0 +1,13 @@
+from flask import url_for
+
+
+def pagination_links(pagination, endpoint, **kwargs):
+    if 'page' in kwargs:
+        kwargs.pop('page', None)
+    links = dict()
+    if pagination.has_prev:
+        links['prev'] = url_for(endpoint, page=pagination.prev_num, **kwargs)
+    if pagination.has_next:
+        links['next'] = url_for(endpoint, page=pagination.next_num, **kwargs)
+        links['last'] = url_for(endpoint, page=pagination.pages, **kwargs)
+    return links
